A HUGE fix, and some other minor changes.
I changed the makefile to be more name consistent.
xpad360w now stands for Xpad 360 Wired.

I made a huge discovery that work structs aren't
meant to be used asyncronously... probably bad to assume that it
was okay to do that in the first place. Major data race which happened
pretty commonly.

HUGE amounts of stability! I'm currently being hardpressed to find bugs.
I need to check input but I'm very pleased with how the driver is working
at this point in time. Coming up pretty quickly, the only thing left to
do will be to reverse engineer some of the announcement messages
and to add more controllers to the USB match macro.