So, this commit just removes input processing all together.
It shows that outside of input processing... everything works perfectly.
In actuality though, I may not have even needed to use mutexes.
As long as I checked for null when processing events, we should
be fine given the in urb only current handles one packet at a time.

However, I haven't determined if allocating a new buffer and asyncronously
offloading the processing is faster or not... none of that matters really
since the processing causes the entire system to hang for some reason anyways.