Several fixes
I fucked up big time on that last commit. If you compile and attempted to ran that, then I'm flat out sorry. Let's not let that happen again.

This feels like the most stable tree I've had in forever though. If anyones looking at this repo, this is the version of the repo you want for now.